Feat: 여행지 평균 평점 추가

+++ src/main/java/coursemaker/coursemaker/domain/review/dto/RequestCourseDto.java
@@ -31,7 +31,7 @@ public class RequestCourseDto {
 
     @Schema(description = "평점", example = "4.5")
     @NotNull(message = "평점을 입력하세요.")
-    private BigDecimal rating;
+    private Double rating;
 
     public CourseReview toEntity(Member member) {
         CourseReview courseReview = new CourseReview();
